improved the resource meta data editor/ inspector helper class
