+++ core/controller/level/ResourceLoader.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using UnityEngine;
 using WorldWizards.core.entity.gameObject.resource;
+using WorldWizards.core.entity.gameObject.resource.metaData;
 
 namespace WorldWizards.core.controller.level
 {
