--- core/entity/gameObject/resource/WWObjectMetaData.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using WorldWizards.core.entity.common;
-
-namespace WorldWizards.core.entity.gameObject.resource
-{
-    [Serializable]
-    public class WWObjectMetaData
-    {
-        public WWType type;
-        public int baseTileSize;
-    }
-}
