+++ core/entity/gameObject/resource/metaData.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 871d96af2add45adaa36cf3bf0962431
+timeCreated: 1511820415
