+++ core/entity/gameObject/resource/metaData/Editor.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 25d6475d54344952af9f2eaa59bf8fa2
+folderAsset: yes
+timeCreated: 1511882803
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
