+++ core/entity/gameObject/resource/metaData/Editor/WWResourceMetaDataEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: bb609eee8410480299da6f84ffa569f7
+timeCreated: 1511820507
