+++ core/entity/gameObject/resource/metaData/WWDoorHolderMetaData.cs.meta
