+++ core/entity/gameObject/resource/metaData/WWDoorMetaData.cs
@@ -2,12 +2,12 @@
 using UnityEngine;
 using WorldWizards.core.entity.common;
 
-namespace WorldWizards.core.entity.gameObject.resource
+namespace WorldWizards.core.entity.gameObject.resource.metaData
 {
     [Serializable]
     public class WWDoor
     {
-        public Vector3 pivot;
+//        public Vector3 pivotOffset;
         public Vector3 facingDirection;
         public float width;
         public float height;
