+++ core/entity/gameObject/resource/metaData/WWResourceMetaData.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using UnityEngine;
+
+namespace WorldWizards.core.entity.gameObject.resource.metaData
+{
+    [Serializable]
+    public class WWResourceMetaData : MonoBehaviour
+    {
+        public WWObjectMetaData wwObjectMetaData;
+        public WWTileMetaData wwTileMetaData;
+        public WWDoor door;
+    }
+}
