+++ core/entity/gameObject/resource/metaData/WWTileMetaData.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 
-namespace WorldWizards.core.entity.gameObject.resource
+namespace WorldWizards.core.entity.gameObject.resource.metaData
 {
     [Serializable]
     public class WWTileMetaData
     {
-        public WWOccupiedWalls wwOccupiedWalls;
+        public WWWallMetaData wwWallMetaData;
         public WWDoorHolder northWwDoorHolder;
         public WWDoorHolder southWwDoorHolder;
         public WWDoorHolder eastWwDoorHolder;
