+++ core/entity/gameObject/utils/WWWallsHelper.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine.Assertions;
 using WorldWizards.core.entity.gameObject.resource;
+using WorldWizards.core.entity.gameObject.resource.metaData;
 
 namespace WorldWizards.core.entity.gameObject.utils
 {
@@ -20,36 +21,36 @@ namespace WorldWizards.core.entity.gameObject.utils
 
             if (yRotation == 0 || yRotation == 360)
             {
-                north = metaData.wwTileMetaData.wwOccupiedWalls.north;
-                east = metaData.wwTileMetaData.wwOccupiedWalls.east;
-                south = metaData.wwTileMetaData.wwOccupiedWalls.south;
-                west = metaData.wwTileMetaData.wwOccupiedWalls.west;
+                north = metaData.wwTileMetaData.wwWallMetaData.north;
+                east = metaData.wwTileMetaData.wwWallMetaData.east;
+                south = metaData.wwTileMetaData.wwWallMetaData.south;
+                west = metaData.wwTileMetaData.wwWallMetaData.west;
             }
             else if (yRotation == 90)
             {
-                north = metaData.wwTileMetaData.wwOccupiedWalls.west;
-                east = metaData.wwTileMetaData.wwOccupiedWalls.north;
-                south = metaData.wwTileMetaData.wwOccupiedWalls.east;
-                west = metaData.wwTileMetaData.wwOccupiedWalls.south;
+                north = metaData.wwTileMetaData.wwWallMetaData.west;
+                east = metaData.wwTileMetaData.wwWallMetaData.north;
+                south = metaData.wwTileMetaData.wwWallMetaData.east;
+                west = metaData.wwTileMetaData.wwWallMetaData.south;
             }
             else if (yRotation == 180)
             {
-                north = metaData.wwTileMetaData.wwOccupiedWalls.south;
-                east = metaData.wwTileMetaData.wwOccupiedWalls.west;
-                south = metaData.wwTileMetaData.wwOccupiedWalls.north;
-                west = metaData.wwTileMetaData.wwOccupiedWalls.east;
+                north = metaData.wwTileMetaData.wwWallMetaData.south;
+                east = metaData.wwTileMetaData.wwWallMetaData.west;
+                south = metaData.wwTileMetaData.wwWallMetaData.north;
+                west = metaData.wwTileMetaData.wwWallMetaData.east;
             }
             else // (yRotation == 270)
             {
-                north = metaData.wwTileMetaData.wwOccupiedWalls.east;
-                east = metaData.wwTileMetaData.wwOccupiedWalls.south;
-                south = metaData.wwTileMetaData.wwOccupiedWalls.west;
-                west = metaData.wwTileMetaData.wwOccupiedWalls.north;
+                north = metaData.wwTileMetaData.wwWallMetaData.east;
+                east = metaData.wwTileMetaData.wwWallMetaData.south;
+                south = metaData.wwTileMetaData.wwWallMetaData.west;
+                west = metaData.wwTileMetaData.wwWallMetaData.north;
             }
 
-            bool top = metaData.wwTileMetaData.wwOccupiedWalls.top;
-            bool bottom = metaData.wwTileMetaData.wwOccupiedWalls.bottom;
-            var rotatedMetaData = new WWOccupiedWalls(north, east, south, west, top, bottom);
+            bool top = metaData.wwTileMetaData.wwWallMetaData.top;
+            bool bottom = metaData.wwTileMetaData.wwWallMetaData.bottom;
+            var rotatedMetaData = new WWWallMetaData(north, east, south, west, top, bottom);
             WWWalls walls = rotatedMetaData.GetWallsEnum();
 
             return walls;
